arm: dts: ls1021a: add 'always-on' property for arch timer
Issue: LIN7-1472

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: arm: dts: ls1021a: add 'always-on' property for arch timer
Date: Fri, 7 Nov 2014 14:07:05 +0800

The arm arch timer is the only available timer on this cpu. Since this
timer is tightly coupled with the core logic and can be lost on core
power down, the CLOCK_EVT_FEAT_C3STOP is set to the arm arm timer event
device. This will disable the high-resolution timer functionality. But
at the moment, we only support the wait mode power management feature
for the core, which means that the power of core never get down,
so add 'always-on' property for this timer to workaround this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>