edac: mpc85xx: don't do the ecc scrubbing for the reserved memory
Issue: LIN7-1419

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: edac: mpc85xx: don't do the ecc scrubbing for the reserved memory
Date: Fri, 14 Nov 2014 10:46:18 +0800

On the mpc85xx platform, we use lwarx/stwcx pair to do a atomic
scrubbing when detecting a single bit ecc error. This method works
fine for normal memory. But for some reserved memories such as the
dedicated private memories for QMAN or BMAN, the cache coherence is
disabled for performance. If a single bit ecc error occurs in these
memory areas, the atomic scrubbing may conflict with the QMAN/BMAN
dma store access and then crash the data there.

In order to fix this issue, we choose to skip all the ecc scrubbing
for the reserved memories. We implement this by using the
ctl_page_to_phys callback of mem_ctl_info struct. If an ecc error
occurs in a reserved memory, we would return a invalid physical
address. This will cause the edac core to skip the ecc scrubbing.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>