rtmutex: drop standalone patch tracking merge fixup for squashed ver.
I'd added a stand-alone commit to track what was done in a merge
commit in the linux-stable-rt fast forward version, when making a
rebased version.  SteveR, when doing the rebased version himself
decided to just fold the contents of the merge fixup into an existing
patch, so we align with that for ease of future maintenance.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>