rt: close on re-sync with rt62.
Aside from the previous merge fixup, rt62 didn't really
bring in any new changes.  We do this just to ensure that
the version numbers and tags match what people would find
in the linux-stable-rt repo.  As of this point, we are matched
with the -rt62, but we keep the local refreshes to:

   net-Use-get_cpu_light-in-ip_send_unicast_reply.patch
   net-ip_send_unicast_reply-add-missing-local-serializ.patch

as those are required for the newer 3.10.60 baseline we have
in use here.  More specifically, we see those as follows:

 -------------------------------
-@@ -1537,7 +1538,7 @@ void ip_send_unicast_reply(struct net *net, struct sk_buff *skb, __be32 daddr,
+@@ -1531,7 +1532,7 @@ void ip_send_unicast_reply(struct net *net, struct sk_buff *skb, __be32 daddr,
  		ip_push_pending_frames(sk, &fl4);
  	}
- out:
+
 -	put_cpu_var(unicast_sock);
 +	put_cpu_light();

  	ip_rt_put(rt);
  }
 -------------------------------

and

 -------------------------------
-@@ -1538,7 +1541,7 @@ void ip_send_unicast_reply(struct net *net, struct sk_buff *skb, __be32 daddr,
+@@ -1532,7 +1535,7 @@ void ip_send_unicast_reply(struct net *net, struct sk_buff *skb, __be32 daddr,
  		ip_push_pending_frames(sk, &fl4);
  	}
- out:
+
 -	put_cpu_light();
 +	put_locked_var(unicast_lock, unicast_sock);

  	ip_rt_put(rt);
  }
 -------------------------------

i.e. two new labels that appear in context from stable commits
that added the labels.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>