ARM: vgic: move vgic behind IPI in dts
Issue: LIN6-8638

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: ARM: vgic: move vgic behind IPI in dts
Date: Wed, 12 Nov 2014 16:55:25 +0800

To make the dts compatible with old kernel.
For example, a kernel which assumes that the index of IPI mask
is 2 and 3 will not work with the dtb in which the index 2
and 3 is now for vgic. So move the vgic behind IPI to make
it backward compatible.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>