Drop "rcu: Reenable interrupt before calling __call_rcu_nocb"
Issue: OVP-2191

This drops the above commit.  The original intention was to avoid
a wake_up call from the RCU code, since the wake_up used non-raw
locking, which would cause an issue when called from atomic contexts.

However, this issue has since been resolved in another fashion; the
RCU code has been converted to the simple wait queue support, which
uses raw locks for wakeups, so this change is not required.

Further, since RCU is such a fundamental subsystem, we can't be
changing the IRQ context under which it runs without a strict
audit of each call site context.  So a revert puts us back to
where we were, without an implicit context change.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>