iwlwifi: configure the LTR
Issue: LIN7-2362

1/1 [
Author: Emmanuel Grumbach
Email: emmanuel.grumbach@intel.com
Subject: iwlwifi: configure the LTR
Date: Tue, 23 Sep 2014 23:02:41 +0300

commit 9180ac50716a097a407c6d7e7e4589754a922260 upstream.

The LTR is the handshake between the device and the root
complex about the latency allowed when the bus exits power
save. This configuration was missing and this led to high
latency in the link power up. The end user could experience
high latency in the network because of this.

Cc: <stable@vger.kernel.org> [3.10+]
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Pengyu Ma <pengyu.ma@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>