misc: drop aio patch ; not applicable to OVP
We have no WR specific commits to this file;  only linux-stable
fixes ; the same is for WR6, but a failed merge in that stream
resulted in this aio patch being present here.

Specifically, in the WR6 fast-forward tree, see this commit:

  commit 46cc4296d728e17eabcef029b1589d5273365890
  Author: Yue Tao <yue.tao@windriver.com>
  Date:   Thu Sep 25 17:25:20 2014 -0700

    Fix the build errors introduced by merging linux stable kernel

    Signed-off-by: Yue Tao <yue.tao@windriver.com>

Since we don't have the original failed merge, drop the
follow-on fixup patch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>