powerpc: Convert some mftb/mftbu into mfspr
Rally: US49096

1/1 [
Author: Scott Wood
Email: scottwood@freescale.com
Subject: powerpc: Convert some mftb/mftbu into mfspr
Date: Tue, 20 Aug 2013 19:33:12 -0500

commit beb2dc0a7a84be003ce54e98b95d65cc66e6e536 upstream

Some CPUs (such as e500v1/v2) don't implement mftb and will take a
trap.  mfspr should work on everything that has a timebase, and is the
preferred instruction according to ISA v2.06.

Currently we get away with mftb on 85xx because the assembler converts
it to mfspr due to -Wa,-me500.  However, that flag has other effects
that are undesireable for certain targets (e.g.  lwsync is converted to
sync), and is hostile to multiplatform kernels.  Thus we would like to
stop setting it for all e500-family builds.

mftb/mftbu instances which are in 85xx code or common code are
converted.  Instances which will never run on 85xx are left alone.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Xulin Sun  <xulin.sun@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>