refresh: autonominated patch refresh for git-am
During treegen, patches are attempted with "git am" and if that fails,
it is tried with reduced context.  If the patch applies with reduced
context, it is logged (path in kernel cache and the new commit ID) to
a file.  Here, we auto refresh the context of those patches with:

~/treegen/tree$ for i in `cat meta/cfg/scratch/refresh |awk '{print $2}'` ; do F=`grep $i meta/cfg/scratch/refresh|awk '{print $1}'` ; F=`echo $F|sed 's,^.*links/,/home/paul/git/,'` ; rm -f 0001-* ; git extract $i ; cat 0001-* > $F ; done

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>