arm64: clean up KGDB Kconfig duplication mess
We had a patch in the kgdb feature adding it to arm64 by using an
upstream commit, but that was not done properly and so the addition
was many lines away from where the original put it.  This opened
the door for the same patch to be re-applied a second time in the
proper place.

Upon merging, the 2nd patch becomes nonsensical, as it just moves
the config line from the incorrect place to the correct place.

Fix all of this up by putting the correct patch in place where it
is 1st applied, and dropping the second application of it entirely.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>