Notification of death of arbitrary processes
1/1 [
Author: Jianchuan Wang
Email: jianchan.wang@windriver.com
Subject: Notification of death of arbitrary processes
Date: Wed, 22 Jan 2014 07:15:59 -0800

This exposes a new feature which may be called to request
notification when an arbitrary process changes state. The
caller specifies a pid, signal number, and event mask, and
when that pid dies, or is stopped, or anything else that
would normally cause a SIGCHLD, the kernel will send the
specified signal to the caller if the event is in the event
mask originally passed down. The siginfo_t struct will
contain the same information as would be included with SIGCHLD.

This is exposed to userspace via the prctl() call with the
PR_DO_NOTIFY_TASK_STATE option.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>