ARM: arch_timer: improve the resolution as sched_clock
Issue: LIN6-9025

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: ARM: arch_timer: improve the resolution as sched_clock
Date: Tue, 23 Dec 2014 06:56:34 +0000

Now sched_clock_mult is calculated by (NSEC_PER_SEC/arch_timer_rate),
this is not accurate enough, because for example:
if there is a timer with frequence 256MHz, we can get ns per cycle
for this frequence is about 3.9ns. But we can only get 3ns.
The resolusion loses about (3.9-3)/3.9 = 23% and it isn't acceptable.

The most noticeable bad thing is that if we run a task with
100% cpu occupied, we can only get about 76% cpu load for
that task in 'top'.

So introduce sched_clock_shift to improve the resolution like
clocks_calc_mult_shift() is doing.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>