KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag - linux - CVE-2014-8480
Issue: LIN5-19850

1/2 [
Author: Paolo Bonzini
Email: pbonzini@redhat.com
Subject: KVM: Emulate multibyte NOP
Date: Thu, 30 May 2013 13:22:39 +0200

commit 103f98ea64a1b0a67d8a1b23070b4db3533db2b8 upstream

This is encountered when booting RHEL5.9 64-bit.  There is another bug
after this one that is not a simple emulation failure, but this one lets
the boot proceed a bit.

Cc: <stable@vger.kernel.org> # 3.9
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Gleb Natapov <gleb@redhat.com>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

2/2 [
Author: Nadav Amit
Email: namit@cs.technion.ac.il
Subject: KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag
Date: Mon, 5 Jan 2015 14:05:44 +0800

commit 3f6f1480d86bf9fc16c160d803ab1d006e3058d5 upstream

The decode phase of the x86 emulator assumes that every instruction with the
ModRM flag, and which can be used with RIP-relative addressing, has either
SrcMem or DstMem.  This is not the case for several instructions - prefetch,
hint-nop and clflush.

Adding SrcMem|NoAccess for prefetch and hint-nop and SrcMem for clflush.

This fixes CVE-2014-8480.

Fixes: 41061cdb98a0bec464278b4db8e894a3121671f5
Cc: stable@vger.kernel.org
Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[YH: 3.4 kernel has no em_clflush function()]
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>