oprofile: grsecurity changes results in x86 page fault
1/1 [
Author: Amy Fong
Email: amy.fong@windriver.com
Subject: oprofile: grsecurity changes results in x86 page fault
Date: Wed, 20 Feb 2013 07:31:35 -0800

This fix is required for x86, an earlier fix for x86-64
was created:

    oprofile: grsecurity changes results in x86-64 page fault

In a nutshell, pax moves the thread info structure off of
the kernel stack and into task struct, so we need to modify
how the information is accessed. Otherwise, the kernel ends
up accessing the wrong information and we eventually end up
with a page fault.

Signed-off-by: Amy Fong <amy.fong@windriver.com>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>