fsl-p50xx: nand oobtest failed
Issue: LIN6-9227

1/2 [
Author: Akinobu Mita
Email: akinobu.mita@gmail.com
Subject: mtd: mtd_oobtest: generate consistent data for verification
Date: Sat, 8 Mar 2014 00:24:10 +0900

commit be54f8f1c76890f4b6163715aed5a3d0a7309dc2 upstream

mtd_oobtest writes OOB, read it back and verify.  The verification is
not correctly done if oobsize is not multiple of 4.  Although the data
to be written and the data to be compared are generated by several
prandom_byte_state() calls starting with the same seed, these two are
generated with the different size and different number of calls.

Due to the implementation of prandom_byte_state() if the size on each
call is not multiple of 4, the resulting data is not always same.

This fixes it by just calling prandom_byte_state() once and using
correct range instead of calling it multiple times for each.

Reported-by: George Cherian <george.cherian@ti.com>
Reported-by: Lothar Waßmann <LW@KARO-electronics.de>
Tested-by: Lothar Waßmann <LW@KARO-electronics.de>
Cc: George Cherian <george.cherian@ti.com>
Cc: Lothar Waßmann <LW@KARO-electronics.de>
Cc: David Woodhouse <dwmw2@infradead.org>
Cc: Lee Jones <lee.jones@linaro.org>
Cc: linux-mtd@lists.infradead.org
Signed-off-by: Akinobu Mita <akinobu.mita@gmail.com>
Signed-off-by: Brian Norris <computersforpeace@gmail.com>
[Make some changes on the original patch which is applied to oobtest.c other
than mtd_oobtest.c.]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

2/2 [
Author: Brian Norris
Email: computersforpeace@gmail.com
Subject: mtd: tests: fix integer overflow issues
Date: Mon, 21 Jul 2014 19:07:12 -0700

commit 1001ff7a4f64f3f4264e69d3ed70ff428f627e01 upstream

These multiplications are done with 32-bit arithmetic, then converted to
64-bit. We should widen the integers first to prevent overflow. This
could be a problem for large (>4GB) MTD's.

Detected by Coverity.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Cc: Akinobu Mita <akinobu.mita@gmail.com>
[Make some changes on the original patch, some files have been removed
the "mtd_" prefix in upstream, and current version doesn't have mtd_test.c]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>