KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag - linux - CVE-2014-8480
Issue: SCP6-526

1/1 [
Author: Nadav Amit
Email: namit@cs.technion.ac.il
Subject: KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag
Date: Mon, 5 Jan 2015 14:14:40 +0800

commit 3f6f1480d86bf9fc16c160d803ab1d006e3058d5 upstream

The decode phase of the x86 emulator assumes that every instruction with the
ModRM flag, and which can be used with RIP-relative addressing, has either
SrcMem or DstMem.  This is not the case for several instructions - prefetch,
hint-nop and clflush.

Adding SrcMem|NoAccess for prefetch and hint-nop and SrcMem for clflush.

This fixes CVE-2014-8480.

Fixes: 41061cdb98a0bec464278b4db8e894a3121671f5
Cc: stable@vger.kernel.org
Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[YH: 3.10 kernel has no em_clflush function()]
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>