intel-iommu: allow ignoring Ethernet device RMRR with IOMMU passthrough
Issue: LIN6-9347

1/2 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: intel-iommu: rewrite RMRR detection to be readable
Date: Wed, 28 Jan 2015 13:12:34 -0500

The original code was confusing because it triggered on a
non match, and it wasn't easily extensible to add in more
devices with RMRR issues.

Rewrite it in a way that is equivalent, but allows more
devices to be listed and instrumentation to be added.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>
]

2/2 [
Author: Matt Peters
Email: matt.peters@windriver.com
Subject: intel-iommu: allow ignoring Ethernet device RMRR with IOMMU passthrough
Date: Mon, 26 Jan 2015 14:25:50 -0500

Some BIOS's are reporting DMAR RMRR entries for Ethernet devices
which is causing problems when PCI passthrough is enabled.  These
devices should be able to use the static identity map since the
host should not be enforcing specific address ranges when IOMMU
passthrough is enabled.

Originally-by: Matt Peters <matt.peters@windriver.com>
[PG: Added bootarg wrapper and documentation entries.]
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>