softirq: align run_do_softirq with linux-stable-rt 3.4 branch
Issue: LIN5-20007

1/1 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: softirq: align run_do_softirq with linux-stable-rt 3.4 branch
Date: Mon, 2 Feb 2015 15:43:10 +0800

commit 96c66e7a993a7bf630523a12d48b59636dbae617 ("softirq: Split softirq
locks") was originally taken from commit 448fc1cab4fa056d60477ee4aac5
in linux-stable-rt.git -- which was based on a 3.8-rt kernel.

If we look at that original linux-stable-rt commit, we see it adds
a run_ksoftirqd() near the top of the patch and removes the existing
run_ksoftirqd() near the bottom of the patch.  However 96c66e7a993a
has no such removal, and so duplicate run_ksoftirqd functions result.
This happens only when RT_FULL is not enabled.

However, if we look at linux-stable-rt v3.4-rt-features branch, we
see that the 3.4.x version of that commit, in commit 3424406bd15ace
does not lead off with the same added definition of run_ksoftirqd
and its helper function ksoftirqd_softirq_pending.  And of course
nor is there the removal at the end of the patch.  Instead only
the helper function is added further down in the patch.

Here we align this 3.4.x based kernel with the linux-stable-rt
v3.4-rt-features version of softirq.c in respect to this difference.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>