sunrpc/cache: allow threads to block while waiting for cache update
Issue: LIN4-32215

1/2 [
Author: J. Bruce Fields
Email: bfields@redhat.com
Subject: nfsd4: allow numeric idmapping
Date: Thu, 22 Mar 2012 16:07:18 -0400

Commit e9541ce8efc22c233a045f091c2b969923709038 upstream.

Mimic the client side by providing a module parameter that turns off
idmapping in the auth_sys case, for backwards compatibility with NFSv2
and NFSv3.

Unlike in the client case, we don't have any way to negotiate, since the
client can return an error to us if it doesn't like the id that we
return to it in (for example) a getattr call.

However, it has always been possible for servers to return numeric id's,
and as far as we're aware clients have always been able to handle them.

Also, in the auth_sys case clients already need to have numeric id's the
same between client and server.

Therefore we believe it's safe to default this to on; but the module
parameter is available to return to previous behavior if this proves to
be a problem in some unexpected setup.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

2/2 [
Author: NeilBrown
Email: neilb@suse.de
Subject: sunrpc/cache: allow threads to block while waiting for cache update.
Date: Thu, 12 Aug 2010 17:04:06 +1000

Commit f16b6e8d838b2e2bb4561201311c66ac02ad67df upstream.

The current practice of waiting for cache updates by queueing the
whole request to be retried has (at least) two problems.

1/ With NFSv4, requests can be quite complex and re-trying a whole
  request when a later part fails should only be a last-resort, not a
  normal practice.

2/ Large requests, and in particular any 'write' request, will not be
  queued by the current code and doing so would be undesirable.

In many cases only a very sort wait is needed before the cache gets
valid data.

So, providing the underlying transport permits it by setting
 ->thread_wait,
arrange to wait briefly for an upcall to be completed (as reflected in
the clearing of CACHE_PENDING).
If the short wait was not long enough and CACHE_PENDING is still set,
fall back on the old approach.

The 'thread_wait' value is set to 5 seconds when there are spare
threads, and 1 second when there are no spare threads.

These values are probably much higher than needed, but will ensure
some forward progress.

Note that as we only request an update for a non-valid item, and as
non-valid items are updated in place it is extremely unlikely that
cache_check will return -ETIMEDOUT.  Normally cache_defer_req will
sleep for a short while and then find that the item is_valid.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>