Revert "wr-rt: add cgroup sleeeping while atomic fix"
This reverts commit 652ee43282e44dff4aba354a69f8af0402bc6428.

I think there is more to this issue than just a lock conversion.
Firstly, if we look at the existing -rt patches, we've got the old
patch from ~2009 that is:

 From: Ingo Molnar <mingo@elte.hu>
 Date: Fri, 3 Jul 2009 08:44:33 -0500
 Subject: [PATCH] core: Do not disable interrupts on RT in res_counter.c

which changed the local_irq_save to local_irq_save_nort in order to
avoid such a raw lock conversion.

Also, when I test this patch on a larger machine with lots of cores, I
get boot up issues (general protection fault while trying to access the
raw lock) or RCU stalls that trigger broadcast NMI backtraces; both
which implicate the same code area, and they go away with a revert.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>