KEYS: close race between key lookup and freeing - linux - CVE-2014-9529
Issue: LIN6-9250

1/1 [
Author: Sasha Levin
Email: sasha.levin@oracle.com
Subject: KEYS: close race between key lookup and freeing
Date: Fri, 27 Feb 2015 10:45:08 +0800

commit a3a8784454692dd72e5d5d34dcdab17b4420e74c upstream

When a key is being garbage collected, it's key->user would get put before
the ->destroy() callback is called, where the key is removed from it's
respective tracking structures.

This leaves a key hanging in a semi-invalid state which leaves a window open
for a different task to try an access key->user. An example is
find_keyring_by_name() which would dereference key->user for a key that is
in the process of being garbage collected (where key->user was freed but
->destroy() wasn't called yet - so it's still present in the linked list).

This would cause either a panic, or corrupt memory.

Fixes CVE-2014-9529.

Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>