MX6: UART 1 hangs the system on SARBESD when not used a console
Issue: LIN5-20069

1/1 [
Author: Lijun Guo
Email: Lijun.Guo@windriver.com
Subject: MX6: UART 1 hangs the system on SARBESD when not used a console
Date: Thu, 12 Feb 2015 13:26:33 +0800

original patch link:
https://git.congatec.com/arm/qmx6_kernel/commit/4a523565f6783c3361b77c6241afae898fc629ef

The platform hangs if UART 1 is not used a console on MX6 SABRESD. The
issue is because the UART UFCR register is not initialized and the
driver uses values provided by the bootloader  (U-Boot in this case).
This patch initializes the UFCR register and does not rely on the
default or what is configured by the bootloader.

Test method on MX6 SABRESD:
	1. Add 'noconsole' to the kernel command line.
	2. SSH to the board and type the below from the SSH window:
	   echo Hello >/dev/ttymxc1
	3. Platform will hang, with the patch the hang will not be seen.

Signed-off-by: Mahesh Mahadevan <Mahesh.Mahadevan@freescale.com>
Signed-off-by: Lijun Guo <Lijun.Guo@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>