Revert "spi: Make max_tx and max_rx the same type"
This reverts commit c723a791fbfebb2a8f05a1add6b87a4f45bd1447.

It is CGP7 specific, and the code it tries to patch is not in
the WR7 baseline.

It tries to patch the spi_map_msg() function, but that does
not exist.  It was added in:

  commit 99adef310f682d6343cb40c1f6c9c25a4b3a450d
  Author: Mark Brown <broonie@linaro.org>
  Date:   Thu Jan 16 12:22:43 2014 +0000

    spi: Provide core support for DMA mapping transfers

That code is not present in 3.14 unless one has a backport:

linux-head$git describe --contains 99adef310f682d6343cb40c1f6c9c25a4b3a450d
v3.15-rc1~147^2~2^2~9
linux-head$

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>