mtd: fsl_ifc_nand: remove incorrect kfree()
Rally: US55336

1/1 [
Author: Brian Norris
Email: computersforpeace@gmail.com
Subject: mtd: fsl_ifc_nand: remove incorrect kfree()
Date: Thu, 28 Mar 2013 21:28:56 -0700

commit b966a40a3854e1263b3528ef95bc9e6017e8e515 upstream

The struct fsl_ifc_mtd is allocated with devm_kzalloc, so its memory
is "managed" automatically by the kernel. That is, we do not need to
free it explicitly; it will be freed when the device is removed. And we
*certainly* shouldn't free it with a regular kfree().

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>