rt: relocate reverts to top of queue of new patches
Ideally we'd put the reverts right beside what they revert, and
then just comment them out, but I tried that and the timer revert
breaks follow on patches.

So for now, just put them together at the top of the new patches added
by -rt32.  If we detect a problem, we can test the reverts independently
of the new patches.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>