stable: fix CVE-2014-8172 and fix fallout in OOT fs/aufs code
Issue: OVP-2357

The mainline commit isn't an ideal candidate for backport ; it removes
functionality used by the Out Of Tree fs/aufs code, and it requires
another commit backported ahead of it.

For the AUFS code, we look at the version for 3.14 which aready has the
mainline commits backported to 3.10 here, present.  A patch is added to
AUFS which makes it look like 3.14 in the areas where the removed
functionality was previously.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>