core2-32: reduce default NR_CPUs to 8
The generic smp fragment sets NR_CPUs to 64, but the Kconfig options for
the core2 processor family limit the max value to 8. This results in
the following warning:

   WARNING: [kernel config]: specified values did not make it into the kernel's final configuration:

   Value requested for CONFIG_NR_CPUS not in final ".config"
   Requested value: "CONFIG_NR_CPUS=64"
   Actual value set: "CONFIG_NR_CPUS=8"

To fix the issue, we make the core2-32 machine set the number of cpus
to a value that it knows are valid, and hence override the general default.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>