cpufreq: drop local workaround patch; original issue resolved differently
Issue: OVP-2390

It was observed that powertop was crashing/failing to start.  This was
traced back to cpufreq files not opening, and returning -EINVAL ; which
was a known RT issue ; see 3.18.9-rt4 release notes.  The complication
here was the locally authored (WR7) patch deleted in this commit.

The -rt originally had this "rwsem-rt-Do-not-allow-readers-to-nest.patch" and
this cpufreq patch from the WR7 team was added to mitigate fallout from it.

On OVP, we integrated "Revert-rwsem-rt-Do-not-allow-readers-to-nest.patch"
from a newer -rt, and so the original issue should have been resolved.
However we still had the original WR7 workaround which now would actively
break things in the allowed-nest environment.

Drop the WR7 workaround and let the original and its revert cancel each
other out, and all is well with the world.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>