bcm-xlp: adjust hyperexec kmod Makefile and Kconfig
1/2 [
Author: Quanyang Wang
Email: quanyang.wang@windriver.com
Subject: bcm-xlp: delete useless configuration in netlogic lib's Makefile
Date: Wed, 29 Apr 2015 15:07:56 +0800

Delete the useless configuration in netlogic lib's Makefile.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
]

2/2 [
Author: Quanyang Wang
Email: quanyang.wang@windriver.com
Subject: bcm-xlp: adjust hyperexec kmod Makefile and Kconfig
Date: Wed, 29 Apr 2015 15:07:56 +0800

Adjust hyperexec kernel module's Makefile and Kconfig for
two reasons. One is because original dependency in Kconfig
and Makefile is not correct and the description is not
identical to SDK's. The other is now we do not support
hyperexec kernel module unless the nae which is network
module, so we make other modules not to built into kernel.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>