neighbour: fix c predence error
Issue: LIN5-19697

1/1 [
Author: yzhu1
Email: yanjun.zhu@windriver.com
Subject: neighbour: fix c predence error
Date: Thu, 30 Apr 2015 13:43:24 +0800

In commit 5331a4"net: neighbour: Add mcast_resolicit to configure
the number of multicast resolicitations in PROBE state", the
intentios is to compute the number of probes. But because of C
predence, the correct number of probes is not got. So the
paranthesis is needed to make sure the correct number.

Reported-by: Slabbert <Clinton.Slabbert@windriver.com>
Signed-off-by: yzhu1 <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>