driver core: Move driver_data back to struct device
Issue: LIN7-1812

1/1 [
Author: Jean Delvare
Email: jdelvare@suse.de
Subject: driver core: Move driver_data back to struct device
Date: Mon, 14 Apr 2014 12:54:47 +0200

commit 1bb6c08abfb653ce6e65d8ab4ddef403227afedf upstream

Having to allocate memory as part of dev_set_drvdata() is a problem
because that memory may never get freed if the device itself is not
created. So move driver_data back to struct device.

This is a partial revert of commit b4028437.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>