powerpc: mpc85xx: flush the cache before cpu down in kexec
commit c97319cc8a4246048cfdf542c12b8b4d48fdc3a0 upstream.

Issue: LIN7-3730

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc: mpc85xx: flush the cache before cpu down in kexec
Date: Mon, 18 May 2015 01:04:15 +0000

We observe a "Zero PT_NOTE entries found" warning when vmcore_init()
is running on the dump-capture kernel. Actually the PT_NOTE segments
is not empty, but the entries generated by crash_save_cpu() are not
flushed to the memory before we reset these cores. So we should flush
the cache as what we do in cpu hotplug. With this change, we can also
kill the mpc85xx_smp_flush_dcache_kexec() since that becomes
unnecessary.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>