bcm-xlp: guest kernel rebase from SDK 3.0.2 to 3.2
Turn on full preemption on the guest. There is a known bug
in SDK 3.2 that the guest hangs with second reboot if
the host loads any kernel module, so the changes in
xlp-ovp-host.cfg build in all kernel modules on the host
as the workaround to fix the bug for now. Also, remove unused
bcm-xlp-add-IP-address-option-to-the-xlp9xx-guest.patch

Signed-off-by: Nam Ninh <nam.ninh@windriver.com>