rt: uprev to latest stable version (with patches!)
New content, not just a stable merge.  Drop obvious overlap with the
irqwork/nohz patch (shows up as a rename) that we backported earlier.

There will clearly be overlap with these as well:

       new file:   features/rt/KVM-lapic-mark-LAPIC-timer-handler-as-irqsafe.patch
       new file:   features/rt/KVM-use-simple-waitqueue-for-vcpu-wq.patch

as we (OVP) have had local variants which did the same thing dating back
to OVP5.  We will deal with those as a separate commit, letting us ensure
that we maintain functional equivalence by comparing the local versions
to the -rt ones when the local ones fail to apply @ treegen time.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>