rt: relocate nohz fix back to rcu dir
This was originally in the rcu dir, since it was in 3.18-rt and to
backport it meant needing irq work changes backported from post
3.14 as well.

It was added by the stable-rt team, but not build tested, so they
didn't notice the dependencies I'd already encountered above.

Issue reported at:
	http://marc.info/?l=linux-rt-users&m=143282614506419&w=2

For now, go back to having our version where it belongs and
leaving the new addition to -rt out of the rt queue.

This is essentially a partial revert of: commit 40f12b8496bf3fc293c
("rt: uprev to latest stable version (with patches!)")

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>